feat: 프로젝트 설정및 valid 해결

+++ src/main/java/dev/momo/api/category/CategoryServiceImpl.java
@@ -3,6 +3,14 @@ package dev.momo.api.category;
 import dev.momo.api.category.dto.CategoryDto;
 import dev.momo.api.category.entity.Category;
 import dev.momo.api.category.repository.CategoryRepository;
+
+import dev.momo.api.global.response.BaseResponse;
+import dev.momo.api.global.response.BaseResponseStatus;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
+
+import dev.momo.api.global.exception.CategoryNotFoundException;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,7 +27,7 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    @Transactional
+    @Transactional //트랜잭션은 데이터가 변경될 경우만 사용하기! 자원 낭비가 된다
     public CategoryDto createCategory(CategoryDto dto) {
         //CategoryEntity setter 안쓰고 저장하는법! -> builder 사용
         Category category = Category.builder()
@@ -36,8 +44,11 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    @Transactional
-    public List<CategoryDto> readAllCategory() {
+
+    public List<CategoryDto> readAllCategory() throws CategoryNotFoundException {
+        if(categoryRepository.findAll().isEmpty())
+            throw new CategoryNotFoundException();
+
         return categoryRepository.findAll().stream()
                 .map(category -> CategoryDto.builder()
                         .categoryId(category.getCategoryId())
@@ -46,9 +57,13 @@ public class CategoryServiceImpl implements CategoryService{
                 .collect(Collectors.toList());
     }
 
+
     @Override
-    public CategoryDto readCategory(Long categoryId) {
+    public CategoryDto readCategory(Long categoryId) throws CategoryNotFoundException {
         Optional<Category> category = categoryRepository.findById(categoryId);
+        if(category.isEmpty())
+            throw new CategoryNotFoundException();
+
         CategoryDto categoryDto = CategoryDto.builder()
                 .categoryId(category.get().getCategoryId())
                 .category(category.get().getCategory())
@@ -57,7 +72,11 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    public boolean updateCategory(Long categoryId, CategoryDto dto) {
+    @Transactional
+    public boolean updateCategory(Long categoryId, CategoryDto dto) throws CategoryNotFoundException {
+        if (!categoryRepository.existsById(categoryId))
+            throw new CategoryNotFoundException();
+
         Optional<Category> category = categoryRepository.findById(categoryId);
         Category category1 = Category.builder()
                 .categoryId(category.get().getCategoryId())
@@ -68,9 +87,25 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    public boolean deleteCategory(Long categoryId) {
+    @Transactional
+    public boolean deleteCategory(Long categoryId) throws CategoryNotFoundException {
+        if (!categoryRepository.existsById(categoryId))
+            throw new CategoryNotFoundException();
+
         Optional<Category> category = categoryRepository.findById(categoryId);
         categoryRepository.delete(category.get());
         return true;
+    }
+
+    //패이징처리용 테스트 용도
+    @Override
+    public List<CategoryDto> readAllCategory1 (Pageable pageable) {
+        List<CategoryDto> list = categoryRepository.findAll(pageable).stream()
+                .map(category -> CategoryDto.builder()
+                        .categoryId(category.getCategoryId())
+                        .category(category.getCategory())
+                        .build())
+                .collect(Collectors.toList());
+    return list;
     }
 }
